refactor(bi): split mnemonics enum into multiple dicts
As per conversation with Stan on 2025-02-20, it will be difficult to teach an IR with an IR servicing both the parser and code generator.
We want to avoid revisiting any listings, so code is effectively append-only.
As such, having multiple maps represent out Mnemonics rather than a mega-class is beneficial.